# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/sources": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "List all sources",
                "operationId": "21bca661d3026520af28d6595d6178fa",
                "responses": {
                    "200": {
                        "description": "List of sources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Source"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sources"
                ],
                "summary": "Create a new source",
                "operationId": "4e849f567869e0a9e93439527982a110",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Source"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Source created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Source"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sources/{id}": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "Get a specific source",
                "operationId": "3c3e2eb8ae0b84260498c0846ad21e0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the source",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Source"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sources"
                ],
                "summary": "Update a source",
                "operationId": "88661edf70ffed04c99dc52a637b8160",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the source",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Source"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Source updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Source"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sources"
                ],
                "summary": "Delete a source",
                "operationId": "d8b9fb302c5aa48e5a7866eb64c7e91f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the source",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Source deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Source": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use 'Bearer {token}' to authenticate requests.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}